Develop

+++ Repositories/RoleRepository.cs
@@ -3,6 +3,8 @@ using RoleService.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.Data.SqlClient;
 
 namespace RoleService.Repositories
 {
@@ -15,28 +17,42 @@ namespace RoleService.Repositories
             _context = context;
         }
 
-        public IEnumerable<Role> GetAll()
+        public Task<List<Role>> GetAll() => Task.FromResult(_context.Roles.ToList());
+
+        public async Task<Role> GetById(Guid id) => await _context.Roles.FindAsync(id);
+
+        public async Task<Role> Create(Role entity)
         {
-            return _context.Roles.ToList();
+            await _context.Roles.AddAsync(entity);
+            await _context.SaveChangesAsync();
+
+            return entity;
         }
 
-        public Role GetById(Guid id)
-        {
-            return _context.Roles.Find(id);
-        }
-
-        public void Create(Role entity)
-        {
-            _context.Roles.Add(entity);
-            _context.SaveChanges();
-        }
-
-        public void Remove(Guid id)
+        public async Task<bool> Remove(Guid id)
         {
             Role role = _context.Roles.Find(id);
+            if (role == null)
+            {
+                return false;
+            }
 
             _context.Roles.Remove(role);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync();
+
+            return true;
+        }
+
+        public async Task<bool> Cleanup()
+        {
+            List<Role> roles = _context.Roles.ToList();
+            for (int i = 0; i < roles.Count; ++i)
+            {
+                _context.Roles.Remove(roles[i]);
+                await _context.SaveChangesAsync();
+            }
+
+            return true;
         }
     }
 }
